Screen/Linux+OSX: Try to reenable display head auto-detection for AMD.
The function to auto-probe and assign screenid -> GPU crtc mappings
was never defective - or so i think. The bug that caused failure was
somewhere else and fixed a year ago, so lets give this function a
second chance to prove its value.

We enable the function on OSX and Linux for AMD gpus again. Also
enhancements to handle Linux multi-x-screen setups.

-> Single and dual display testing on Linux + AMD with single X-Screen
   (one or two active outputs) and dual X-Screen (one output per screen)
   looks good, multi-display tests and OSX tests pending.