Screen/Windows: Add HiDPI (trouble) detection code.
Try to find out about our host process DPI awareness level and the
effective DPI of displays on which we display. Factoring in the Windows
version (Vista+? Win-8+?), DWM state, single- vs. multi-display, global
and per monitor DPI, decide if and what kind of info/warning messages
should be given to user in case of potential or guaranteed HiDPI + DWM
induced visual stimulus onset timing trouble.

Due to the specific shortcomings of Windows in this area this is less
than fool-proof and it is not clear of how much value the
implementation is. Also only tested on Win-7 VM so far…