Enable PsychOculusVRCore for Octave-4 on Windows.
This is a weird one. Copy the Oculus runtime DLL from the SysWOW64
folder into a location private to us, then link against that instead of
libOVR, as libOVR as provided by the Oculus 0.5.0.1 SDK is incompatible
with MinGW, but the runtime DLL is happily accepted by the GNU linker.