PR655init: Do flush on write on OSX.
Change flushing behaviour after writes for the PR655 on OSX. We now
flush after writes. Forum message 19808 suggests this improves
reliability at least on OSX 10.10. Linux otoh. requires not flushing
according to older messages and the status of Windows is unknown, so we
stick to the old Linux behaviour of not flushing.

I keep the other PR driver for 650/700 etc. unmodified for the time
being, as this may be different on each OS, or potentially based on the
connection type or on the serial port driver used.

Time will tell if this is a good change of settings.