Screen: Fix GStreamer on Octave-4 for Windows.
The official 32-Bit Octave-4.0.0 upstream build for Windows includes a
version of libbz2.dll which is incompatible with the libbz2.dll that is
bundled with GStreamer 1.6.0. This causes important GStreamer plugins
to fail to load. Solve this by overriding Octave’s libbz2.dll with
GStreamers dll. We do this during GStreamer init by LoadLibraryEx()
loading of that dll from the GStreamer runtime directory, whose path we
get communicated via a environment variable set by the PsychStartup.m
startup script.

If that environment variable would not be set we would use the standard
Octave dll, so just copying a suitable dll into the Octave/bin folder,
e.g., from the GStreamer bin folder, would also work.

Replacing Octave’s dll is probably safe: Octave includes the dll to
support image loading of compressed MIFF (Magick image file format)
image files via GraphicsMagick - the engine behind imread(). So the
override may either have no negative effect at all, assuming our “good”
dll is backwards compatible, or it would likely be limited to the
inability to imread() compressed MIFF image files. I have not ever seen
such a file during regular use and don’t expect any to be ever used on
Octave for Windows, so this is an acceptable tradeoff.

With this change we still get two “could not load plugin” warnings,
because Octave also ships a libstdc++-6.dll incompatible with
GStreamers variant. This causes openh264dec.dll and libgsttaglib.dll to
fail loading - luckily they are not crucial to our needs. We can’t
replace libstc++, as its exported symbols are incompatible with the
GStreamer variant — not just an omission but major incompatibility, so
Octave would not start anymore when replacing this dll.

All in all this is a bit shaky, as it is unclear if other important
dll’s would be incompatible. Otoh., testing of movie playback and movie
writing shows that both work reliably.

What currently does not work is video capture. Octave terminates
immediately in VideoCaptureDemo.m. Reason unknown: The debugger shows
no crash, but apparently exit() or return() calls of Octave being
responsible for shutdown. I couldn’t attach any breakpoint to get any
backtraces, so the reason for the termination is completely unclear.
However, video capture is not the most critical and most widely used
functionality, especially not on MS-Windows, so this is ok for a first
release. Lets first find out if anybody wants or needs Octave for
Windows anyway.

-> PsychStartup.m updated to export an env variable with the gstreamer
runtime directory for use/pickup by Screen.mex.