Screen/Linux: Add a new specialflag kPsychSafeForDRI3.
This windowRecord special flag is set to signal that Screen thinks
the use of DRI3/Present is safe on the given setup, ie., that the
X-Server is recent enough (1.16.3 or later), and Mesa is recent
enough (10.5.2 or later). It is used to signal DRI3 suitability
to usercode via the 'GetWindowInfo' SpecialFlags field. Mostly
used by to be committed setup code.