Screen/Windows: First try to workaround new DWM madness in Windows 10.
This patch is based on first underwhelming tests on Windows 10.

Apparently the DWM compositor never goes to standby completely
on a Windows-10 system. Since Windows 8 we are unable to force
the compositor off, so have to trust the DWM to auto-disable as
needed and appropriate, e.g., for opaque, top-level, unoccluded,
decorationless fullscreen windows when no HiDPI scaling is needed.

First testing on Windows-10 with a Radeon HD-5770 single-display
and dual-display setup suggests the DWM never goes completely to
standby as hoped. At least our hacks to indirectly detect if it
does so - and which used to work on Windows-7 at some time - no
always report the DWM active. Either it really stays on all the time
or our hacks got broken by changes to Windows. It is unknown if the
same is already true on Windows 8 or 8.1, but we don't care about
that systems in the first place, as users can upgrade to Windows-10
if they want.

So now while some of our adjustments for compositor delay on Win-10
seem to be still needed - the ones for refresh calibration, the ones
we do for swap scheduling now seem to be counterproductive and cause
failure instead of fixing it.

Therefore restrict compositor delay adjustment to < Windows-8 systems,
or in the case of Windows 8 or later, restrict it to video refresh
calibration, but exclude swap scheduling adjustments.

-> Rebuild and test of Screen pending.
-> This is just a guess based on first results, probably needs
   more iterations, or is entirely wrong.