DaqAInScan: Use interfaceID to define IndexRange or scans.
Try to select interfaces ascending by interfaceID for assignment to
IndexRange, instead of hard-coding ranges. This should not change
anything on Linux or Windows, but maybe it helps the brain-dead OSX.

-> Should not change anything on Linux or Windows which both work well
for DaqAInScan so far, but may help OSX. Ok, probably not, but it can’t
get more broken on OSX than now.
-> Test by users pending.