Screen/Linux/X11: Clear backbuffer to black and do a final flip at window close.
This to make sure we switch back from a fullscreen window to the
windowing system with a well defined solid black frontbuffer.
On some multi-x-screen setups, where the non-zero stimulation
X-Screen is not managed by the WM, the X-Screen continues to
display whatever the last frontbuffer image was at onscreen
window close time. We'd rather have it display a solid eye
and monitor friendly black to avoid burn in, than displaying
whatever the last stimulus frame was.