XOrgConfCreator: Maybe fix it for NVidia proprietary driver.
The NVidia blob does not recognize Option 'ZaphodHeads', but instead
calls it 'UseDisplayDevice' to use that on nvidia ddx. However the
driver also does not seem to use the RandR names of outputs for that
option and i don't know atm. how to enumerate its names, so this
will probably not work either. Anyway, worth a try in the wild...