Eyelink demos cleanup.
God is this awful!

Don't use commandwindow without checking if that function is
actually supported, like on Octave-3. Don't use exception
handling when there are methods that are more compatible to
Octave-3.

And FFS don't use "clear all" at the beginning of the
body of the main function!!!

I'm stopping after cleaning up 4 demos because otherwise
i might puke.