ListenChar(): Add new flag -1 to only suppress spilling into console.
This suppresses keystrokes from showing up in the Matlab/Octave
command window or console, just as a flag of 2, but it doesn't
enable listening for those characters.

Iow. this is for pure input suppression when other methods like
KbCheck and keyboard queues are used for keyboard input.

Using only ListenChar(-1) - ListenChar(0) pairs allows to get
the suppression without the need to engage keyboard queues in
configurations where we usually would need to use them, e.g.,
MS-Windows Vista and later, Matlab in -nojvm or -nodesktop mode,
or Octave with/without GUI.

This allows to more easily avoid trouble with mixing ListenChar
and keyboard queues.

The downside is that without keyboard queues, we can not detect
CTRL+C to reenable keyboard input to Octave or to Matlab in -nojvm
or -nodesktop mode, so the user code must be more careful to have
proper error handling and reenable stuff properly via ListenChar(0).