Screen/OSX: Make affine DrawText transforms suck less on Apple Coretext.
The implementation of affine text transforms with Apples CoreText
legacy text renderer 0 now sucks only half as much as before, but still
enough to be an essentially pointless pain to use.

If we detect a non-identity text transform we now switch to a slow mode
of operation where we allocate a Quartz CGBitmapContext the full size
of the target window, rasterise text into that context, then try to
only get roughly the bits onto the screen which sort of matter. This
sort of works, and makes text transforms slightly less brain damaged,
but still sucks:

- Performance breaks down to a fraction of normal performance.

- The returned text bounding box has almost nothing to do with the real
location of the text.

- Some new OSX memory management bugs get triggered and cause the
application process to leak memory like hell, accumulating dozens of GB
of virtual memory quickly, eventually crashing due to out of memory.

-> So this is mostly useless except for really masochistic people. We
tried to make it work, we failed, time to move on.