DrawFormattedText: Be more clever in bounding box calculation.
At least try to. This approach seems to work ok with the
FTGL text renderer, ie., our standard text renderer plugin
for all OS'es, the level of brokeness on the legacy renderers
is unknown.

This is supposed to compute tighter bounding boxes for
multi-line text, especially with a non-default vSpacing > 1.

See PTB forum message 20240 for reference of the problem
we try to solve here.