DrawFormattedText(): Add sx='centerblock', always use yPosIsBaseline, update help text.
We used yPositionIsBasline inconsistently. Used it if some sx
horizontal formatting option was given, but not if an absolute
sx or none was given. Adding to the weirdness, why let horizontal
formatting options decide about how vertical text positioning is
done?

-> Now we always use yPositionIsBaseline. for y positioning of text.

Add new option sx == 'centerblock' which horizontally centers
a block of text in the window. The block of text defined by the
text string is centered as a whole, shifting it so the longest
line of text is centered, instead of centering each line of text
in the string individually.

-> Update help text to also stress the approximative nature of
   text alignment and bounding box calculations.