New demo: TurnTableDemo.m
TurnTableDemo uses PsychPortAudio's mixing and streaming
refill functions to simulate a multi-channel audio mixer
and a record turntable like DJ's use it for some rapping
and scratching. A turntable is simulated which can be
free-running, or driven manually by mouse movement,
touchpad/touchscreen movement, or by a connected Griffin
PowerMate dial knob. Buttons allow to enable/disable
additional audio sample loops which get mixed with the
main track from the turn table (Sample loops and batteries
not included in package).

This demo was sitting for 3 years on my harddisc, waiting
for its moment to step into the light. Initially it was
meant to demo PTB's tablet features, but as tablet support
for the Nexus-7 was not a big success, this demo wasn't
ever released.

Now with Griffin PowerMate knob support, this demo is a
cute way to demo the knob. The knob, if attached and
configured (currently only under Linux), allows to
control the virtual turntable of this demo to do some
serious scratching.

-> This demo is arguably one of the more pointless ones,
   but it is too cool to be omitted, and it shows some
   advanced use of PsychPortAudio, so there...