Provide zip compressed DownloadPsychtoolbox.m.zip
Just DownloadPsychtoolbox.m zip compressed.

The idea, as suggested by Denis, is to link to this file from the
Download section of the PTB Wiki, because this way a simple click onto
the link might get users a proper file on their disc drive, instead of
requiring them to right-click- and-save. Apparently some people are
challenged in how to use a web browser in the year 2016 and this could
help. We will see…