PsychHID/Windows: In HIDAPI increase max number of reports to 512
WindowsXP and later support storing up to 512 HID input reports in the
kernel. Rebuild HIDAPI for Windows to request this maximum number of
slots, instead of the upstream setting of 64 reports.

-> This to make handling of the Griffin PowerMate a bit more robust.
One “click” of rotational knob movement queues 1 report, so at 512
slots we can queue up to 512 clicks to store about 5 full rotations,
whereas the old 64 slots would not even allow a subject to turn the
knob a full rotation without the driver losing data.