PsychHID/Windows: Also enumerate custom devices for KbQueues
Not only what can be classified clearly as keyboard, mouse or joystick.
However, this is not complete, we need special setup code for each
custom device, which we currently don’t have, so this is a bit
pointless atm.

It does enumerate the PowerMate, but can’t make use of it then.