BitsPlusPlus: Add new subfunction 'UseFE1StereoGoggles'
This allows to configure the PTB imaging pipeline to
drive CRS FE1 stereo goggles connected to a Bits
display device. The pipeline will draw proper
T-Lock codes to the bottom of each stimulus
frame to set the goggles for the following
frame.

The old CRS toolbox helper function bitsGoggles was
updated to compute proper T-Lock codes and return
them to BitsPlusPlus instead of directly executing
them.

BitsStereoDemo.m was updated to use modern PTB
concepts and demonstrate the goggles with the Bits
devices.

-> Test pending. Just coded, not tested.