PsychGPUControl: Add Unity+Compiz multi-display workaround.
Due to the design (or a design limitation, depending on ones
viewpoint) of Compiz fullscreen window unredirection, the
compositor only unredirects windows which are "fullscreen" on
any given Compiz logical output, *not* on a X-Screen.

By default Compiz auto-detects outputs and their geometry via
RandR, ie. 1 RandR output == 1 Compiz output.

This means that for a single display setup, the only detected
RandR output corresponds to the full X-Screen, so a window
that is fullscreen on the X-Screen is therefore fullscreen on
the single output and Compiz unredirects the window properly.

On a X-Screen with multiple outputs attached, ergo multiple
RandR outputs == multiple Compiz outputs, each output only
covers part of a X-Screen. Therefore a fullscreen window
will not exactly cover one single output but multiple outputs,
therefore Compiz does not unredirect the window and we won't
get rid of compositing, which is obviously bad for our visual
timing.

The proper solution would be to create a patch for Compiz merged
upstream and then integrated into current Linux distros, so
that Compiz not only unredirects windows which are "fullscreen"
on any single RandR output == Compiz output, but also fullscreen
on the complete X-Screen - covering all outputs.

Our current workaround is to disable auto-detection of outputs
by Compiz and instead manually define a "list" with exactly 1
output by changing some Compiz config settings in the Unity
profile via dconf. This one output is super-sized 50000x50000
pixels, way larger than any real display x-screen. Compiz will
automatically clamp the size of such too large outputs to the
true size of the X-Screen, therefore Compiz will end up with
a single output corresponding exactly to the geometry of the
X-Screen. Therefore our fullscreen window on the X-Screen
will also exactly cover the single Compiz output and Compiz will
unredirect it, so it can use page-flipping for good visual
stimulation timing.

We integrate the setup code into a new PsychGPUControl's new
subfunction:

PsychGPUControl('EnableCompizMultiDisplayWorkaround', enableFlag);

Sometimes the compositor seems to not pick up the new settings
immediately, so a logout -> login cycle of the user may be
neccessary.

Downsides:

* Works with no known problems on a single-display setup.

* Works with no known problems on a multi-display multi-x-screen
  setup, if the stimulation displays are all on secondary x-screens
  ie. not x-screen 0, as Compiz does not care about such secondary
  x-screens, so never redirects windows.

* On a single-x-screen, dual-display setup, the auto-positioning/
  sizing of regular GUI windows is a bit off wrt. the optimal
  ergonomic behavior for multi-display, because Compiz seems to use
  the geometry of its logical outputs to decide about window placement.
  Because with our workaround the compositor no longer sees two outputs
  on a dual-display setup with exactly the geometry of the two
  displays, maximizing a window on one monitor no longer works -
  the window now maximizes over the whole desktop. Dragging a
  window to the left or right edge of the desktop no longer
  positions and resizes the window to fill exactly the left or
  right half of a display, but instead the left or right half
  of the whole desktop. This is obviously not the optimal ergonomic
  behavior but unavoidable when our hack is in use in order to
  fix multi-display single-x-screen visual stimulation for stereo
  dual-display setups.

-> Clearly the Compiz/Unity default behavior is designed with the
   intention of a desktop with ergonomic window placement that
   at the same time optimizes display performance for typical
   fullscreen desktop applications like slide presentation software,
   video players, or video games, as unredirection will still provide
   increased performance of such apps. So the behavior is by design.

   It's just unfortunate for us that the special case of fullscreen
   on multiple displays applications didn't get optimized and that
   i didn't get to the root cause a few month earlier to submit an
   upstream fix in time for the LTS release. Now Ubuntu 16.04-LTS is
   released and chances are probably that we will be stuck with this
   problem and hack for most/all of the 16.04 LTS cycle.