Screen('DrawTextures'): Optimize batch drawing performance.
If a batch of to be drawn primitives is drawn with one single
'filterMode' and from one single source texture, ie. only the
other properties of the batch are potentially per primitive,
then use an optimized new helper routine for drawing of the
batch with the minimal number of GL state changes:

PsychBatchBlitTexturesToDisplay() gets called to start a batch,
telling it the total number of primitives in the batch, then
for each to be drawn primitive, then a final time to finalize
the batch. The routine tries to avoid state changes as much as
possible by setting up and tearing down state only once per
batch if possible.

Current limitations:

- Still uses immediate mode primitive submission.

Results:

- Works fairly well.

- CPU overhead is reduced by a factor of 3x - 4x. This shifts some
  workloads from being completely CPU bound to being completely GPU
  bound, at least for the given test case on a fast 16-core Intel
  Xeon X5550 at 2.67 GhZ, with a Radeon HD 5770 under Linux Ubuntu
  16.04.0 LTS + Mesa 11.2.0 + radeon-kms, with default power management
  settings.

  E.g., Speedup for test from 200 fps to 700 fps with vsync disabled
  under DRI3/Present (aka use of non-vsync'ed copy-swaps) on a full
  HD 1920x1080 display.