Screen/Linux: Add a workaround for potential VC4 Mesa DRI3/Present bug.
Mesa's Gallium VC4 driver doesn't like it at all if we
perform OpenGL bufferswaps without rendering anything
inbetween the swaps. Doing so during video refresh
calibration causes permanent failure of visual stimulation,
ie. we see massive flicker and pixel trash, because the
system effectively turns into a single-buffered mode where
the backbuffer == frontbuffer, ie. the render target buffer
is the scanout buffer, and we can see the rendering happening
in all its ugly glory.

This is the same problem i observed on Intel kms under DRI3/Present.
So far i thought it is an Intel ddx sna bug, but now i think it is
a bug in Mesa's DRI3/Present backend. Something goes wrong with the
renderbuffer assignment logic and the logic gets stuck on only handing
out one fixed renderbuffer which is also the scanout buffer. The problem
needs proper investigation and fixing in Mesa.

Our old workaround was to always render the startup splash image or
at least clear the framebuffer to white inbetween swaps. This works
on other drivers, but VC4 is too slow with our standard PsychDrawSplashImage,
which is why i made it skip that drawing on VC4, which is why the workaround
didn't work around. Now upon detection of VC4, we draw a single pixel into
the framebuffer - touching the framebuffer to work around the bug, but a
cheap enough operation to prevent timing trouble during video refresh
calibration and sync tests.

-> With this hack applied, GLSLDemo, ShepardZoomDemo and PerceptualVBLSyncTest
   now work without flicker, rendering corruption etc and debug code in the
   vc4 kms driver shows that Mesa is properly alternating between different
   BO's instead of submitting always the same BO for scanout.