LoadIdentityClut: Add detection and handling of Broadcom VC4.
The VideoCore-4 in RapsberryPi needs gfx lut type 0.
Luckily type 0 was the default used whenever a gpu
was unknown, so even before today PTB worked flawlessly
with VC4 :). This addition is just to replace the
"unknown gpu, the sky might be falling" message with
a friendly "relax, all is good" message.