Screen: Only glFlush() the pipeline on AsyncFlipBegin, do not glFinish().
glFinish must be used for frame-sequential stereo mode 11 to make
sure blocking for render completion happens in the main interpreter
thread, and not in the stereo flipper thread.

For pure async flip however, a glFlush() is sufficient before
unsetting the OpenGL context on the main interpreter thread.
This avoids needless stalling of the main thread - and thereby
Matlab/Octave M-Code - by needless synchronization to the GPU.

-> This makes a huge difference in performance for certain
   Async-Flipped workloads, e.g., PsychProPixx() 4x and 12x
   modes for 480 Hz or 1440 Hz redraw rate when using flipmode 1
   for async flips.

-> Tested successfully on Linux, rebuilt of Screen on main build
   system pending.