PsychGPUTestAndTweakGammaTables(): Replace 'WaitBlanking' by simple WaitSecs.
Screen('WaitBlanking') turns out to be broken on some graphics
drivers + OS combos due to graphics driver bugs (in case of
NVidia proprietary graphics driver 361 for Linux when the
imaging pipeline is in use), or severe OS deficiencies as
on OSX with Intel gpu's.

The primary path via beampos queries doesn't work under skipsyntests 2,
or on Intel gpu's, or on OSX if the PsychtoolboxKernelDriver isn't
used.

The secondary path works on Linux in general, but is broken on
NVidia proprietary drivers of at least the 361 series when the
imaging pipeline is in use. It is always completely dysfunctional
on OSX and Windows.

The 3rd and final fallback path, using a brute force approach,
had a bug when used with the imaging pipeline.

This bug in the 3rd path is fixed now for upcoming PTB 3.0.13 but
i won't fix it in the mex files for 3.0.12 final. Therefore, as an
easily backportable fix, instead get rid of our only important use
for WaitBlanking in PsychGPUTestAndTweakGammaTables(). We can easily
replace it by a simple WaitSecs for a few refresh durations, as this
does not need to be exact wrt. scanout, just provide sufficiently long
pauses for gamma tables and scanout to stabilize to new steady state.

Screen('Waitblanking' is still used in a few ancient CRS demos, but
those are rarely used by anyone, as we demo much better approaches
in more prominent demos wrt. Bits+ and other CRS devices.