BitsPlusImagingPipelineTest: Fix bug in readbackoffset adaptation code.
Since Screen('GetImage') no longer allows placing the samplingRect
(partially) outside the window, this code was broken. Trying a
corrective offset of dx = -1 or dy = -1 would cause the rectangle
to fall outside the legal area and cause an error abort.

Fix this by shifting the "test stimulus" by one unit horizontally
and vertically, and adapting the "readbackoffset (0,0)" location
acccordingly for backbuffer readback. This way we have enough
wiggle room in all directions to shift the readbackoffset again
if needed to compensate for graphics drivers with off-by-one
positioning bugs.

This was broken since January 2016, and now reported by Justin Ales
in forum message #20719. Hopefully this will fix the problem.