Screen/Linux: Add some more roundoff fudge for OpenML correctness test.
A 1 msec fudge in one of the code-path was missing. This can
cause a false positive reporting of swap scheduling failure
in some pathological corner cases.

In the past only observed on RaspberryPi. Now Justin Ales
reported the corner case on a 9 year old iMac with Radeon
Mobility HD 2400 XT gpu (RV 610 / M74), so we take this
more seriously. The RPi would have needed this extra fudge
anyway, as its time stamping won't be guaranteed to be
more accurate than a maximum error of 1 vblank duration,
ie. ~1 msecs.