Add support for basic control of RaspberryPi's GPIO pins.
RPiGPIOMex() is a new mex file to control the Pi's GPIO
pins with the help of the wiringPi library. It is meant
as a stop-gap solution until the IOPort NG driver is
brought into existence.

linuxmakeitoctave3(13) builds the mex file, requiring
the wiringPi libraries and headers.

RaspberryPiGPIODemo.m shows how to use it. It demos how to:

- Wait for triggers on a pin efficiently, using RPi's trigger irqs.
- Get/Set digital pin state.
- For demo purpose, the RPi 2B's two builtin LEDs are controlled,
  this will only work on 2nd gen Pi's and maybe the new 3rd gen?

-> PsychLinuxConfiguration.m now also adds users on ARM/RPi
   to the Unix user group 'gpio' for non-root access to the
   gpio pins.