Screen/Linux: Add hybrid graphics laptop detection support/handling.
Use the pci_is_boot_vga() function of libpciaccess to find out which
of multiple gpu's is the actual display gpu and try to attach to that
gpu, so we attach to the right gpu for low-level access on dual-gpu
laptops etc. Also return the id of the display gpu in a new field
of the struct returned by Screen('GetWindowInfo') so M-Files can
detect the display gpu vs. render gpu. Needed for LoadIdentityClut
to select linearized gamma tables for the display gpu in such a laptop.

If we detect that DRI PRIME renderoffload is requested via the
DRI_PRIME environment variable, we are more lenient wrt. framerate
during refresh calibration/sync tests. We also allow for half the
expected flip rate, as some hybrid graphics muxless "Optimus" and
"Enduro" laptop gpu combos are just too slow to swap at full video
refresh rate.

As a bonus, now allow to map all gpus read-only except AMD, as a
read-only mapping is fully sufficient on those. Only AMD can actually
benefit from a read-write mapping.

-> Screen for 64-Bit Octave on Linux rebuilt and lightly tested.
   More tests pending.