Can't get PTB root in a script with nested functions
In Octave/MATLAB, you can't add a variable to the workspace of
a function that contains a nested function. This breaks the function
PsychRuntimeGetPsychtoolboxRoot because it tries to assign the
Psychtoolbox root directory to a global variable. When that fails,
the mex doesn't know PTB's root directory and fails to load the text
renderer plugin, for example.

Note that this fix uses mexSetTrapFlag() that MathWorks may deprecate
in the near future. As of this commit, Octave doesn't support the
function meant to replace mexSetTrapFlag's functionality.