PsychLinuxConfiguration: Fix setup of AMD deep color support.
Also trigger a update-initramfs after user enabled deep color
support. The AMD kms drivers reside in the initrd, so it is
not enough to update the modprobe config files for deep color,
but one needs to update the initramfs image, so these options
get applied to the drivers at system boot time.