"help HybridGraphics": Update to latest state for amdgpu-kms
The root cause for the sync problems on Intel + AMD VI class
has been found: amdgpu-kms uses shared dmabuf fences on the
reservation object of the exported dmabuf, instead of an
exclusive fence. Fixing that would likely fix the prime sync
failure. The work still needs to be done though.

I have a patch that hacks the intel driver to wait for all
fences instead, which fixes the problem now, but is not
acceptable for upstream as it is a hack, not a proper
solution.

-> Update help text wrt. scope of affected AMD gpus (Volcanic
   Islands only) and timeframe for fix.

-> Also update recommendation to use Linux 4.7.3 instead of
   Linux 4.7.2 at this time for the recommended kernel.