DownloadLegacyPsychtoolbox.m: It's dead, Jim!
The GoogleCode website is dead, a mere placeholder and
redirect to our GitHub site, so all the url's are wrong.

Moreover we lost all branches and version tags execpt
master when we exported from GoogleCode, so all reference
points for getting the code for one of the legacy versions
are gone.

This needs a complete overhaul in how we deal with legacy
versions. Or if we deal with them at all. In the last couple
of years, only 1 person ever asked for a legacy version
download, so maybe not worth the hassle.