Update PsychEyelinkDispatchCallback.m
The isempty(3) is a typo, I assume this wants to check if a PTB window is open before proceeding. Perhaps it should use `Screen(windowPtr, 'WindowKind')` but I'm not sure if this would add overhead to the callback, and checking if it is empty should do the same thing as the only place `eyewin` is set (line 112) that check has already been done.