DaqAInScan: Replace deprecated bitcmp in 12 bit handling.
In Matlab R2013a backwards compatibility for bitcmp was
broken, because idiots at work. Replace the use of bitcmp
for handling of 12 bit 2s-complement encoded numbers by
something more simple and hopefully tamper proof against
idiots. Fix the likely wrong threshold of > 2048 by >= 2048
for detecting the negative 12th sign bit being set.

Strangely this 12 bit complement encoding only seems to
be used for differential channels, not single ended channels.
Is this correct? It definitely is weird. In any case no hw
to test, so lets just hope for the best.