Screen/Linux: Add detection and handling of NVidia Optimus on proprietary driver.
Detect if PTB is running on a DRI PRIME Sync setup "NVidia proprietary driver style":

The X-Screen master GPU (typically NVidia dGPU) is running as output source,
the X-Screen slave GPU (typically Intel iGPU) is running as output slave.
Synchronization is done vie double-buffering/page-flipping, dmabuf-fences
and the new handshaking implemented by NVidia in XOrg 1.19 + modesetting ddx.

Iow. all rendering for the X-Screen, 2D and 3D/OpenGL is done by the
NVidia output source gpu, the output sink gpu only provides display.
This is different from the DRI3/Present renderoffload configuration
used by the open-source graphics driver.

We detect this hybrid graphics configuration by checking for the
presence of a RandR output property called "PRIME Synchronization"
which gets only created by the X-Server on such a setup.

The design description, as implemented in XOrg-1.19-rc and the
NVidia 370.23 beta driver, can be found in Andy Ritgers XDC 2016
talk here (slides and video of the talk):

https://www.x.org/wiki/Events/XDC2016/Program/ritget_prime/

The problem with this approach is that we essentially have no accurate
control over visual stimulus onset timing _and_ no way of getting
reliable and trustworthy visual stimulus onset timestamps. Also we
have no idea of when actual vblanks happen on the display gpu. This
makes it mostly impossible to provide research grade visual timing
or timestamping with the current implementation as of XOrg 1.19 and
NVidia 370 series.

So what do we do if we detect such an Optimus + Proprietary driver setup?

a) Notify/Warn user about the bad timing properties and lack of reliable
   timing on the setup. Advice the user to switch to the open-source
   nouveau graphics/display driver instead if research grade timing is
   needed. This part works.

b) Try to compensate for the problems, to make the magnitude of the
   errors less worse - but not really good. This part may be hopeless.
   The current implementation seems to make timestamps less wrong, but
   still pretty bad, but then it seems to make visual stimulus onset
   timing less stable/more erratic.

   The driver can be in one of two possible states as far as i understand
   the talk. Either the system is idle. Then a swapbuffers request will
   be executed without extra lag, ie., potentially at the next vblank if
   the system is fast enough. However then the timestamps will be meaning-
   less/highly inaccurate. Or the system is in a steady state of constant
   animation, in which case the timestamps might be approximately correct,
   but the swap scheduling incurs 1 frame of lag.

   After rewatching Andy's talk i think this might be a no-win scenario,
   where we can't trick the system into behaving ok'ish, just wrong in
   a different way. Anyway, more reading/analysis of X-Server code, potential
   tinkering with the modesetting ddx driver part of the system, and testing
   is needed to see if we can improve something or not, so this just sets
   some baseline to modify and tinker with.

-> Screen for 64-Bit Octave/Matlab on Linux rebuilt and lightly tested.