Screen/Linux: Add workaround for Mesa 12 bug in multi-x-screen configs.
Workaround for a bug introduced into Mesa 12.0, potentially backported to some 11.2 releases,
by commit cf804b4455fac9e585b3600a8318caaced9c23de. glXCreateNewContext() does not perform
some validation of fbconfig correctly against the X-Screen stored in the fbconfig for our
target X-Screen, but instead always validates against X-Screen 0 -- the DefaultScreen(dpy).

This ends badly whenever we are not trying to create a context on X-Screen 0. Work around
this by temporarily assigning dpy->default_screen as our target screen 'scrnum' while
glXCreateNewContext is called. Note: A bug fix for this bug was submitted upstream to resolve
this properly, but that doesn't help us if we need to run on affected Mesa 12.0.x releases,
e.g., as shipping with Ubuntu 16.10, and presumably future Ubuntu 16.04.2-LTS graphics stack
which will be based on the 16.10 stack.

-> Screen.mex for 64-Bit Octave on Linux rebuilt, test pending.