Screen/Linux: Add custom swap completion protocol on Optimus+NVidia blob.
With the Optimus implementation in XOrg 1.19 + the proprietary NVidia
graphics/display driver there is no way we can get reliable flip
completion timestamping, so this calls for a big hack to get a somewhat
reliable and accurate solution:

Implement a custom UDP communication protocol between the XOrg 1.19
modesetting ddx that drives the slaveOutput iGPU - typically an Intel
onboard graphics chip - and Psychtoolbox Screen() mex file.

1. Whenever the modesetting ddx in Prime-Synced outputSlave mode
   either schedules or completes a kms pageflip to update the display
   with new rendered content from the outputSource X-Screen master gpu
   (e.g., the NVidia dGPU + proprietary driver), it sends out a UDP
   packet via loopback to localhost port (10000 + x-ScreenId). This packet
   contains (msc, ust) of the vblank in which the flip either gets scheduled
   or completed + screenId + if this is a flip scheduling or completion.

   A scheduled flip is reasonably expected to complete at the next vblank,
   ie. at (msc + 1, ust + 1 video refresh duration). This is a usually
   correct prediction of flip completion 1 frame ahead of actual completion.

   A completed flip is obviously completed, so a reliable notification
   of flip completion.

2. Screen's PsychOSGetSwapCompletionTimestamp() function receives the UDP
   packets on the corresponding UDP port. Ie. for a fullscreen onscreen
   window on X-Screen screenId it listens on localhost:10000+screenId for
   flip scheduled or flip completed UDP packets from the modesetting ddx.
   This blocks the thread that executes the OpenGL bufferswap until either
   confirmed swap completion or prediction that the swap will complete at
   the next vblank. Then it parses and returns the appropriate timestamps
   for Screen('Flip').

The implementation is only used for fullscreen windows: One problem is that
the ddx doesn't know why a flip gets triggered, and therefore we don't know
if the reason for a completion is our bufferswap, or some visual change due
to other events, e.g., onscreen popup notifications, software cursor updates
etc. To reduce the likelyhood of misattribution the mechanism only works for
fullscreen windows. To reduce the likelyhood of misattributions in fullscreen
mode due to external flip triggers, e.g., mouse cursor updates which are done
by the NVidia blob via a software cursor and therefore trigger updates themselves,
we emit a "swap triggered" packet to ourselves before glXSwapBuffers() and
wait first for reception of that packet, before we wait for the flip packets
from the ddx. This "fence" allows to filter out flip events from the ddx which
were not caused by our own swaps, but external actors.

Another problem is that the current XOrg 1.19 implementation causes a lag
of 1 extra frame between glXSwapbuffers() and flip completion, as the kms
pageflip is only scheduled at the vblank after glXSwapbuffers (best case
scenario) and ergo can only complete one vblank later. This means if we
wait for true flip completion (aka UDP flip completion packet) then that
will only allow Screen('Flip') to achieve a swap every 2nd video refresh,
throttling the max framerate to half the video refresh rate of the display.
To avoid this lag we use the "flip scheduled" UDP packets instead and return
not after swap completion, but 1 frame earlier. This means that the predicted
flip completion timestamps are not 100% reliable, although they will be right
most of the time. This needs some more thought on how to get the best possible
compromise between reliability and latency - possibly a tweakable setting?

-> Current implementation seems to work reliably even during cursor movements.

-> Some light testing with Datapixx confirms that predictive mode provides
   correct and accurate "completion" timestamps. The full stress test is
   still tbd.

-> Performance testing shows we can get full framerate in this op-mode.

-> Test of use of completion events instead of scheduled events is tbd.
-> Various tweaks and improvements are tbd.